refactor: auth 관련 파일에서 apiClient 및 스키마 경로를 lib에서 auth로 변경
