+++ omechu-app/src/app/auth/forgot-password/page.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import { useRouter } from "next/navigation";
 
 import Toast from "@/components/common/Toast";
-import type { FindPasswordFormValues } from "@/lib/schemas/auth.schema";
+import type { FindPasswordFormValues } from "@/auth/schemas/auth.schema";
 
 import ForgotPasswordForm from "./components/ForgotPasswordForm";
