+++ omechu-app/src/app/auth/sign-up/components/TermsAgreement.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { useFormContext } from "react-hook-form";
 
 import Checkbox from "@/auth/components/Checkbox";
-import type { SignupFormValues } from "@/lib/schemas/auth.schema";
+import type { SignupFormValues } from "@/auth/schemas/auth.schema";
 
 type ModalType = "service" | "privacy" | "location";
